Choose class for ClairApi
Conflicts:
	paclair/config_reader.py
